Move version utilities into Util class
